% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_area.R
\name{fisharea}
\alias{fisharea}
\title{Classify coordinates in fishery administration areas}
\usage{
fisharea(x, lon_col = "lon", lat_col = "lat", region = shape_files$id)
}
\arguments{
\item{x}{Data frame containing columns with coordinates.}

\item{lon_col}{Name of longitude column in data frame.}

\item{lat_col}{Name of latitude column in data frame.}

\item{region}{Name of classification system to return names for. Possible
values are "ICES areas", "ICES statistical subrectangles",
"ICES statistical rectangles", "ICES ecoregions", and "NAFO divisions".}
}
\description{
Find which fishery administration area a set of coordinates
lies within to.
}
\examples{
p <- data.frame(lat = c(65.3, 40.0),
                lon = c(-39.3, -9.5))
fisharea(p)
#> [1] "14.b.2" "9.a"

fisharea(p, region = "ICES statistical rectangles")
#> [1] "59B0" "08E0"
}
