% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_area.R
\name{list_to_vector_convert_na}
\alias{list_to_vector_convert_na}
\title{Convert list to vector and replace character(0) with NAs.}
\usage{
list_to_vector_convert_na(l)
}
\arguments{
\item{l}{List to convert}
}
\description{
Convert list to vector and replace character(0) with NAs.
}
