% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_area.R
\name{fish_area}
\alias{fish_area}
\title{Locate point in Fishery administration areas}
\usage{
fish_area(x)
}
\arguments{
\item{x}{Data frame with columns 'lat' and 'long'.}
}
\description{
Locate point in Fishery administration areas
}
