% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_area.R
\name{fisharea_sf}
\alias{fisharea_sf}
\title{Get shape file for fishery administration areas}
\usage{
fisharea_sf(region = shape_files$id)
}
\arguments{
\item{region}{Name of classification system to return name for.}
}
\description{
Get shape file for fishery administration areas
}
